improves order_items pages design

+++ db/migrate/20190215094717_create_products.rb
@@ -4,7 +4,7 @@ class CreateProducts < ActiveRecord::Migration[5.2]
       t.string :title
       t.text :description
       t.string :brand
-      t.decimal :price#, precision: 5, scale: 0, default: 0
+      t.integer :price#, precision: 5, scale: 0, default: 0
       t.integer :quantity
 
       t.timestamps
